+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Next version
+  - Added improved syntax coercion [#32](https://github.com/logstash-plugins/logstash-input-snmp/pull/32)
+
 ## 0.1.0.beta5
   - Added OPAQUE type coercion [#29](https://github.com/logstash-plugins/logstash-input-snmp/pull/29)
   - Added SNMPv3 support [#27](https://github.com/logstash-plugins/logstash-input-snmp/pull/27)